kleaf: Add hermetic_tools.
This change adds a hermetic_tools rule that combines
tools from the host and tools from prebuilts, and
provides a path and a setup script for clients to use.

This change also transforms the following:
  //build/kernel:host-tools
  //build/kernel:kernel-build-scripts , the symlinks to prebuilts
to //build/kernel:hermetic-tools, so we have a central place
of tools to build the kernel.

This change also fixes a bug that the kernel build is not
using host tools from Bazel, but rather its own HERMETIC_TOOLCHAIN.
Previously, when _setup_env.sh sees HERMETIC_TOOLCHAIN=1, it
creates its own list of symlinks to the host-tools, making
//build/kernel:host-tools ineffective. This change introduces
a KLEAF_HERMETIC_TOOLCHAIN flag that kernel.bzl passes to
_setup_env.sh, so it disables the creation of $OUT_DIR/host_tools.
The hermetic_tools rule already sets PATH appropriately.

Test: TH
Bug: 211695803

+++ _setup_env.sh
@@ -99,22 +99,24 @@ BUILDTOOLS_PREBUILT_BIN
 )
 
 if [ "${HERMETIC_TOOLCHAIN:-0}" -eq 1 ]; then
-  HOST_TOOLS=${OUT_DIR}/host_tools
-  rm -rf ${HOST_TOOLS}
-  mkdir -p ${HOST_TOOLS}
-  for tool in \
-      bash \
-      git \
-      install \
-      perl \
-      rsync \
-      sh \
-      tar \
-      ${ADDITIONAL_HOST_TOOLS}
-  do
-      ln -sf $(which $tool) ${HOST_TOOLS}
-  done
-  PATH=${HOST_TOOLS}
+  if [ "${KLEAF_HERMETIC_TOOLCHAIN:-0}" -eq 0 ]; then
+    HOST_TOOLS=${OUT_DIR}/host_tools
+    rm -rf ${HOST_TOOLS}
+    mkdir -p ${HOST_TOOLS}
+    for tool in \
+        bash \
+        git \
+        install \
+        perl \
+        rsync \
+        sh \
+        tar \
+        ${ADDITIONAL_HOST_TOOLS}
+    do
+        ln -sf $(which $tool) ${HOST_TOOLS}
+    done
+    PATH=${HOST_TOOLS}
+  fi
 
   # use relative paths for file name references in the binaries
   # (e.g. debug info)
